Search bar function and profile page alignment changes

+++ js pages/profile.js
@@ -1,6 +1,6 @@
 import { initializeApp } from "https://www.gstatic.com/firebasejs/10.14.1/firebase-app.js";
-import { getFirestore, setDoc, doc, getDocs, getDoc, addDoc, collection, query, where, deleteDoc,updateDoc} from "https://www.gstatic.com/firebasejs/10.14.1/firebase-firestore.js";
-import { getStorage, ref, uploadBytes, getDownloadURL,deleteObject } from "https://www.gstatic.com/firebasejs/10.14.1/firebase-storage.js";
+import { getFirestore, setDoc, doc, getDocs, getDoc, addDoc, collection, query, where, deleteDoc, updateDoc } from "https://www.gstatic.com/firebasejs/10.14.1/firebase-firestore.js";
+import { getStorage, ref, uploadBytes, getDownloadURL, deleteObject } from "https://www.gstatic.com/firebasejs/10.14.1/firebase-storage.js";
 import { getAuth, signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/10.14.1/firebase-auth.js";
 
 // Your web app's Firebase configuration
@@ -21,11 +21,11 @@ const storage = getStorage(app);
 const auth = getAuth(app);
 
 const showImg = document.getElementById("ProfileImg");
-const docRef = doc(db, "users",localStorage.getItem("uid"));
+const docRef = doc(db, "users", localStorage.getItem("uid"));
 const docSnap = await getDoc(docRef);
 console.log(docSnap.data().profileimg);
 
-showImg.src=docSnap.data().profileimg;
+showImg.src = docSnap.data().profileimg;
 
 if (localStorage.getItem("src")) {
     const showImg = document.getElementById("ProfileImg");
@@ -35,27 +35,27 @@ if (localStorage.getItem("src")) {
 
 
 // fetch data from firestore
-async function fetchDatas(){
-    const userPostContainer=document.getElementById("fileDisplay");
-    const querySnapshot = await getDocs (query(collection(db, "Posts"),where("uid","==",localStorage.getItem("uid")) ));
+async function fetchDatas() {
+    const userPostContainer = document.getElementById("fileDisplay");
+    const querySnapshot = await getDocs(query(collection(db, "Posts"), where("uid", "==", localStorage.getItem("uid"))));
 
     console.log(querySnapshot)
 
-        querySnapshot.forEach((doc) => {
-            
-            console.log("Document Data:", doc.data());
+    querySnapshot.forEach((doc) => {
 
-            // Create and configure the image element
-            const imgElement = document.createElement("img");
-            imgElement.classList.add("post-image");
-            imgElement.src = doc.data().postURL;
-            
+        console.log("Document Data:", doc.data());
 
+        // Create and configure the image element
+        const imgElement = document.createElement("img");
+        imgElement.classList.add("post-image");
+        imgElement.src = doc.data().postURL;
 
 
-            // Append the image element to the container
-            userPostContainer.appendChild(imgElement);
-        });
+
+
+        // Append the image element to the container
+        userPostContainer.appendChild(imgElement);
+    });
 
 }
 
@@ -64,11 +64,11 @@ fetchDatas()
 
 // upload pofile from firestore
 const upload = document.getElementById("Choose");
-upload.addEventListener("change", () =>{
+upload.addEventListener("change", () => {
     uploadImage();
 })
 
-async function uploadImage(){
+async function uploadImage() {
 
     const uid = localStorage.getItem("uid");
     const messageShow = document.getElementById("ProfileImg");
@@ -79,33 +79,33 @@ async function uploadImage(){
             const storageRef = ref(storage, `image/${file.name}`);
             await uploadBytes(storageRef, file);
             console.log("File Uploaded Successfully");
-            
-        
+
+
             const downloadURL = await getDownloadURL(storageRef);
-            const docRef = doc(db, "users",localStorage.getItem("uid"));
+            const docRef = doc(db, "users", localStorage.getItem("uid"));
             const docSnap = await getDoc(docRef);
-            await updateDoc(docRef, { profileimg:downloadURL});
+            await updateDoc(docRef, { profileimg: downloadURL });
             messageShow.style.display = "block";
             window.location.reload(true);
-            
+
         } catch (error) {
             console.error("Error:", error);
         }
-    
+
     } else {
         alert("Please upload a valid image file.");
     }
-    
+
 }
 
 ////////////////////////////////////// user name change function /////////////
 const pen = document.getElementById("pen");
-pen.addEventListener("click", () =>{
+pen.addEventListener("click", () => {
     const popup = document.getElementsByClassName("popUp")[0];
-    if(popup.style.display =="none"){
-     popup.style.display ="block";
+    if (popup.style.display == "none") {
+        popup.style.display = "block";
     }
-    else{
+    else {
         popup.style.display = "none"
     }
 });
@@ -115,34 +115,34 @@ pen.addEventListener("click", () =>{
 const okay = document.getElementById("okay");
 const cancel = document.getElementById("cancel");
 const profileXmark = document.getElementById("profileXmark");
-okay.addEventListener("click", (event) =>{
+okay.addEventListener("click", (event) => {
     event.preventDefault();
-    if(valid()){
-    const usernameChange = document.getElementById("usernameChange");
-    const usernameChangeVal = usernameChange.value.trim();
-    const docRef = doc(db, "users",localStorage.getItem("uid"));
-    const docSnap = getDoc(docRef);
-    updateDoc(docRef, {username :usernameChangeVal});
+    if (valid()) {
+        const usernameChange = document.getElementById("usernameChange");
+        const usernameChangeVal = usernameChange.value.trim();
+        const docRef = doc(db, "users", localStorage.getItem("uid"));
+        const docSnap = getDoc(docRef);
+        updateDoc(docRef, { username: usernameChangeVal });
 
-    setTimeout(() => {
-        window.location.reload(true); // Force reload from the server
-    }, 2000);
+        setTimeout(() => {
+            window.location.reload(true); // Force reload from the server
+        }, 2000);
     }
 
 });
 
-profileXmark.addEventListener("click", () =>{
+profileXmark.addEventListener("click", () => {
     const popup = document.getElementsByClassName("popUp")[0];
     popup.style.display = "none";
 })
 
-cancel.addEventListener("click", () =>{
+cancel.addEventListener("click", () => {
     const usernameChange = document.getElementById("usernameChange")
     const usernameChangeVal = usernameChange.value = "";
 })
 ///////////////////////////////////// username valitation //////////////////////
 function valid() {
-        const usernameRegex = /^(?=.*[a-zA-Z])[a-zA-Z][a-zA-Z0-9_-]{3,19}$/
+    const usernameRegex = /^(?=.*[a-zA-Z])[a-zA-Z][a-zA-Z0-9_-]{3,19}$/
 
     const usernameChange = document.getElementById("usernameChange");
     const usernameChangeVal = usernameChange.value.trim();
@@ -158,7 +158,7 @@ function valid() {
         changeError[0].textContent = "length between 3 and 19 char.";
         isValid = false;
     }
-    
+
     else if (!usernameRegex.test(usernameChangeVal)) {
         changeError[0].textContent = "Please enter a valid username";
         isValid = false;
@@ -174,36 +174,62 @@ function valid() {
 
 //ellipsis button action 
 const profileSettings = document.getElementById("ProfileImg");
-profileSettings.addEventListener("click", () =>{
+profileSettings.addEventListener("click", () => {
     const settings = document.getElementsByClassName("settings");
-    if(settings[0].style.display === "none"){
+    if (settings[0].style.display === "none") {
         settings[0].style.display = "block";
     }
-    else{
+    else {
         settings[0].style.display = "none";
     }
 })
 
 
 
-// remove button action
+// Remove button action
 const Remove = document.getElementById("Remove");
-Remove.addEventListener("click", () =>{
-    alert("Are you sure remove the profile");
-    const showImg = document.getElementById("ProfileImg");
-    showImg.removeAttribute("scr");
-    showImg.setAttribute("src", "../assests/photos/dummy-image.jpg") 
+Remove.addEventListener("click", async () => {
+    // Confirm with the user
+    const confirmation = confirm("Are you sure you want to remove the profile?");
+    if (!confirmation) {
+        return; // Exit if the user cancels
+    }
 
-    const fileRef = ref(storage, 'profile/'+profileName);   //delete from storage
-    deleteObject(fileRef)
-    .then(() => {
-      console.log("File deleted successfully");
-    })
-    .catch((error) => {
-      console.error("Error deleting file:", error);
-    });
-    showImg.setAttribute("src", "../assests/photos/dummy-image.jpg") 
-})
+    const showImg = document.getElementById("ProfileImg");
+
+    try {
+        // Get the document reference
+        const docRef = doc(db, "users", localStorage.getItem("uid"));
+        const docSnap = await getDoc(docRef);
+
+        if (docSnap.exists()) {
+            const userData = docSnap.data();
+            const userProfile = userData.profileimg;
+            const fileRef = ref(storage, userProfile); // Reference the file in storage
+
+            // Delete the file from Firebase Storage
+            await deleteObject(fileRef);
+
+            // Update Firestore with the dummy image path
+            const dummyImagePath = "../assests/photos/dummy-image.jpg";
+            await updateDoc(docRef, { profileimg: dummyImagePath });
+
+            // Update the image source in the UI
+            showImg.setAttribute("src", dummyImagePath);
+        } else {
+            console.error("No such document!");
+        }
+
+        console.log("File deleted successfully and profile updated.");
+    } catch (error) {
+        console.error("Error deleting file:", error);
+
+        // Handle errors gracefully
+        alert("Failed to remove the file from storage.");
+    }
+});
+
+
 
 
 // user name fetch function
@@ -258,7 +284,7 @@ const LogoutPost = document.getElementById("ProfileLogout");
 LogoutPost.addEventListener('click', () => {
     // Show the confirmation box before signing out
     const userChoice = confirm("Are you sure you want to log out?");
-    
+
     if (userChoice) {
         // Proceed with signing out
         auth.signOut()
